Level now uses array of function pointers to update, much better. Also touched up Player explosion animation.

+++ inc/enemies.h
@@ -114,7 +114,7 @@ void en1_changeVelocity(Enemy* enemy, Action_Arg vel_x, Action_Arg vel_y, Action
 
 void en1_changePosition(Enemy* enemy, Action_Arg pos_x, Action_Arg pos_y, Action_Arg dir, Action_Arg zerospeed, Action_Arg arg5);
 
-void en1_spawnEnemy(Enemies* enemies, const u16 type, s16 pos_x, s16 pos_y, const Behaviour* behaviour);
+void en1_spawnEnemy(Enemies* enemies, const u8 type, s16 pos_x, s16 pos_y, const Behaviour* behaviour);
 
 void en_chgaction(Enemy* enemy, Action_Arg arg0, Action_Arg arg1, Action_Arg arg2, Action_Arg arg3, Action_Arg arg4, Action_Arg arg5);
